@@ -4,22 +4,26 @@
   "private": true,
   "scripts": {
     "dev": "next dev",
-    "build": "next build && next export",
+    "build": "next build",
     "start": "next start",
     "deploy": "yarn build && gh-pages --dist out --no-history --dotfiles"
   },
   "dependencies": {
-    "@material-ui/core": "^4.12.4",
-    "fontsource-roboto": "^4.0.0",
-    "next": "13.4.2",
-    "react": "18.2.0",
+    "@emotion/cache": "^11.14.0",
+    "@emotion/react": "^11.14.0",
+    "@emotion/server": "^11.11.0",
+    "@emotion/styled": "^11.14.0",
+    "@mui/material": "^7.0.1",
+    "@mui/material-nextjs": "^7.0.0",
+    "next": "15.2.4",
+    "react": "19.1.0",
     "react-copy-to-clipboard": "^5.1.0",
-    "react-dom": "18.2.0"
+    "react-dom": "19.1.0"
   },
   "devDependencies": {
-    "@types/node": "^20.2.1",
-    "@types/react": "^18.2.6",
-    "gh-pages": "^5.0.0",
-    "typescript": "^5.0.4"
+    "@types/node": "^22.13.14",
+    "@types/react": "^19.0.12",
+    "gh-pages": "^6.3.0",
+    "typescript": "^5.8.2"
   }
 }
