@@ -1,43 +1,26 @@
 import * as React from 'react';
-import Document, { Html, Head, Main, NextScript } from 'next/document'
-import { ServerStyleSheets } from "@material-ui/core";
+import { Html, Head, Main, NextScript } from 'next/document'
+import {
+    DocumentHeadTags,
+    documentGetInitialProps,
+} from '@mui/material-nextjs/v15-pagesRouter';
 
-export default class MyDocument extends Document {
-  static async getInitialProps(ctx) {
-    const initialProps = await Document.getInitialProps(ctx)
-    return { ...initialProps }
-  }
 
-  render() {
-    return (
-      <Html lang="en" itemScope itemType="http://schema.org/WebPage" prefix="og: http://ogp.me/ns#">
-        <Head>
-          <script async src="https://inform.everyone.wtf/legal.min.js" data-site-id="0822e697-d6b8-4d0c-9eff-7aa2f3365df1"></script>
-        </Head>
-        <body>
-          <Main />
-          <NextScript />
-        </body>
-      </Html>
-    )
-  }
-};
+export default function Document(props) {
+  return (
+    <Html lang="en" itemScope itemType="http://schema.org/WebPage" prefix="og: http://ogp.me/ns#">
+      <Head>
+        <script async src="https://inform.everyone.wtf/legal.min.js" data-site-id="0822e697-d6b8-4d0c-9eff-7aa2f3365df1"></script>
+      </Head>
+      <body>
+        <Main />
+        <NextScript />
+      </body>
+    </Html>
+  )
+}
 
-MyDocument.getInitialProps = async (ctx) => {
-  // Render app and page and get the context of the page with collected side effects.
-  const sheets = new ServerStyleSheets();
-  const originalRenderPage = ctx.renderPage;
-
-  ctx.renderPage = () =>
-    originalRenderPage({
-      enhanceApp: (App) => (props) => sheets.collect(<App {...props} />),
-    });
-
-  const initialProps = await Document.getInitialProps(ctx);
-
-  return {
-    ...initialProps,
-    // Styles fragment is rendered after the app and page rendering finish.
-    styles: [...React.Children.toArray(initialProps.styles), sheets.getStyleElement()],
-  };
+Document.getInitialProps = async (ctx) => {
+  const finalProps = await documentGetInitialProps(ctx);
+  return finalProps;
 };
