@@ -1,4 +1,10 @@
-import 'fontsource-roboto';
+import { AppCacheProvider } from '@mui/material-nextjs/v15-pagesRouter'
+import type { AppProps } from 'next/app'
 
-import App from "next/app";
-export default App;
+export default function MyApp({ Component, pageProps }: AppProps) {
+   return (
+    <AppCacheProvider>
+        <Component {...pageProps} />
+    </AppCacheProvider>
+   );
+ }
